pyproject.toml
[project]
name = "aidevs4-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.28.0",
    "fastapi>=0.116.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
